will connect main page and card component
